fix return value of CL_cURL_CallbackWrite (#3804)

+++ code/client/cl_curl.c
@@ -218,7 +218,7 @@ static int CL_cURL_CallbackProgress( void *dummy, double dltotal, double dlnow,
 	return 0;
 }
 
-static int CL_cURL_CallbackWrite(void *buffer, size_t size, size_t nmemb,
+static size_t CL_cURL_CallbackWrite(void *buffer, size_t size, size_t nmemb,
 	void *stream)
 {
 	FS_Write( buffer, size*nmemb, ((fileHandle_t*)stream)[0] );
